Setting customer has been created

+++ CheckInn/Model/Room.cs
@@ -10,7 +10,7 @@ namespace CheckInn.Model
     {
         public int RoomID { get; set; }
         public string RoomType { get; set; }
-        public float PricePerNight { get; set; }
+        public decimal PricePerNight { get; set; }
         public string Status { get; set; }
        
     }
